requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==0.8.11
requests==2.31.0
python-dotenv==1.0.0
joblib==1.3.2
torch==2.0.1
transformers==4.31.0
tqdm==4.65.0 